Share .well-known other uses on per-alias-document-root sites.
This stops linking /home/public/{Alias}/.well-known to
/home/public/.well-known.  Instead, it creates /home/public/{Alias}/.well-known
as a directory, and then links acme-challenge inside of it.

It will abort if there is a non-link acme-challenge inside an alias's
.well-known directory, because dehydrated won't be able to work with that.

Credit to github users "quad" and "mina86" who proposed varying fixes for
this issue.  We took a bit of inspiration from both approaches.